requirements.txt
frozendict
python-dateutil
tinydb